:pencil2: Keep spelling error fixed and revert last changes
Reverted last changes and simply changed our UPS password credentials.
We got 'malformed xml document error' due to special characters in our
ups password.

The SOAP library uses xmlEscape on the DATA in the BODY
component, but not DATA in the HEAD component. This is different from
the xml-builder library which uses xmlEscape on both HEAD and BODY.

This meant that we use xmlEscape on every instance use of the SOAP
library. It was getting complicated and error prone. So I decided to
revert the changes.